+++ hockey_scraper/nhl/scrape_functions.py
@@ -54,8 +54,6 @@ def print_errors():
             print(x[0], x[1])
             errors = ' '.join([errors, str(x[0]), ","])
 
-    print('\n')
-
     # Clear them all out for the next call
     game_scraper.broken_shifts_games = []
     game_scraper.broken_pbp_games = []
@@ -102,6 +100,43 @@ def scrape_list_of_games(games, if_scrape_shifts):
     return pbp_df, shifts_df
 
 
+def scrape_schedule(from_date, to_date, data_format='pandas', rescrape=False, docs_dir=False):
+    """
+    Scrape the games schedule in a given range.
+    
+    :param from_date: date you want to scrape from
+    :param to_date: date you want to scrape to 
+    :param data_format: format you want data in - csv or  pandas (pandas is default)
+    :param rescrape: If you want to rescrape pages already scraped. Only applies if you supply a docs dir. (def. = None)
+    :param docs_dir: Directory that either contains previously scraped docs or one that you want them to be deposited 
+                     in after scraping. When True it'll refer to (or if needed create) such a repository in the home
+                     directory. When provided a string it'll try to use that. Here it must be a valid directory otheriwse
+                     it won't work (I won't make it for you). When False the files won't be saved.
+    
+    :return: DataFrame of None
+    """
+    cols = ["game_id", "date", "venue", "home_team", "away_team", "start_time", "home_score", "away_score", "status"]
+
+    # First check if the inputs are good
+    shared.check_data_format(data_format)
+    shared.check_valid_dates(from_date, to_date)
+
+    # Check on the docs_dir and re_scrape
+    shared.add_dir(docs_dir)
+    shared.if_rescrape(rescrape)
+
+    print("Scraping the schedule between {} and {}".format(from_date, to_date))
+
+    # live = True allows us to scrape games that aren't final
+    sched = json_schedule.scrape_schedule(from_date, to_date, preseason=True, not_over=True)
+    sched_df = pd.DataFrame(sched, columns=cols)
+
+    if data_format.lower() == 'csv':
+        shared.to_csv(from_date + '--' + to_date, sched_df, "nhl", "schedule")
+    else:
+        return sched_df
+
+
 def scrape_date_range(from_date, to_date, if_scrape_shifts, data_format='csv', preseason=False, rescrape=False,
                       docs_dir=False):
     """
@@ -133,7 +168,8 @@ def scrape_date_range(from_date, to_date, if_scrape_shifts, data_format='csv', p
     pbp_df, shifts_df = scrape_list_of_games(games, if_scrape_shifts)
 
     if data_format.lower() == 'csv':
-        shared.to_csv(from_date + '--' + to_date, pbp_df, shifts_df, "nhl")
+        shared.to_csv(from_date + '--' + to_date, pbp_df, "nhl", "pbp")
+        shared.to_csv(from_date + '--' + to_date, shifts_df, "nhl", "shifts")
     else:
         return {"pbp": pbp_df, "shifts": shifts_df, "errors": errors} if if_scrape_shifts else {"pbp": pbp_df,
                                                                                                 "errors": errors}
@@ -174,7 +210,8 @@ def scrape_seasons(seasons, if_scrape_shifts, data_format='csv', preseason=False
         pbp_df, shifts_df = scrape_list_of_games(games, if_scrape_shifts)
 
         if data_format.lower() == 'csv':
-            shared.to_csv(str(season) + str(season + 1), pbp_df, shifts_df, "nhl")
+            shared.to_csv(str(season) + str(season + 1), pbp_df, "nhl", "pbp")
+            shared.to_csv(str(season) + str(season + 1), shifts_df, "nhl", "shifts")
         else:
             master_pbps.append(pbp_df)
             master_shifts.append(shifts_df)
@@ -215,7 +252,8 @@ def scrape_games(games, if_scrape_shifts, data_format='csv', rescrape=False, doc
     pbp_df, shifts_df = scrape_list_of_games(games_list, if_scrape_shifts)
 
     if data_format.lower() == 'csv':
-        shared.to_csv(str(random.randint(1, 101)), pbp_df, shifts_df, "nhl")
+        shared.to_csv(str(random.randint(1, 101)), pbp_df, "nhl", "pbp")
+        shared.to_csv(str(random.randint(1, 101)), shifts_df, "nhl", "shifts")
     else:
         return {"pbp": pbp_df, "shifts": shifts_df, "errors": errors} if if_scrape_shifts else {"pbp": pbp_df,
                                                                                                 "errors": errors}